// client/models/UpdateRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateRequestDto = {
    description?: string;
    requestingEntityId?: number;
    targetEntityId?: number;
    contractId?: number;
    mineId?: number;
    materialId?: number;
    startDate?: string;
    endDate?: string;
};